Improved recursive building; Minor fixes in code

+++ modules/tools/include/adefs.h
@@ -11,8 +11,9 @@
 
 namespace tools
 {
-   typedef double Real;
-   typedef double Radian;
+   typedef float Real;
+   typedef float Radian;
+
    typedef std::complex<Real> Complex;
 
    inline Real realAbs(const Real &x) { return ((x) < 0.0)? (-x) : (x); }
